==> Machine Learning/4 Knn.ipynb <==
import pandas as pd
import numpy as np

FILE_TRAIN = 'data/pp-training.csv'
FILE_TEST = 'data/pp-testing.csv'

from sklearn.neighbors import KNeighborsRegressor
from sklearn.model_selection import train_test_split

df = pd.read_csv(FILE_TRAIN, low_memory=False)

allcols = df.columns
alldata = df
X_train, X_test, y_train, y_test = train_test_split(alldata.drop(['price'], axis=1),
                                                    alldata.price, test_size=0.1)


regr = KNeighborsRegressor(n_neighbors=8, algorithm='kd_tree', metric='manhattan', n_jobs=-1)
regr.fit(X_train, y_train)
regr.score(X_test, y_test)

regr = KNeighborsRegressor(n_neighbors=8, algorithm='ball_tree', metric='manhattan', n_jobs=-1)
regr.fit(X_train, y_train)
regr.score(X_test, y_test)

regr = KNeighborsRegressor(metric='euclidean' ,n_neighbors=20, algorithm='kd_tree', weights='distance', n_jobs=-1)
regr.fit(X_train, y_train)
regr.score(X_test, y_test)

cols = list(X_train)

df = pd.read_csv(FILE_TEST, low_memory=False, usecols=cols)
df = df[cols]
predicted_2 = regr.predict(df)

ids = pd.read_csv('data/properati_dataset_testing_noprice.csv')
ids = ids[['id']]
df2 = pd.DataFrame(predicted_2)
df3 = ids.join(df2)
df3.columns = ['id','price_usd']
df3.to_csv('data/prediccion-kaggle-Knn.csv', index=False)
df3 = df3.set_index('id')

df3
